# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CertFlow 企业内部 PKI 证书生命周期管理平台后端 API，提供认证、CA、证书、CRL、OCSP、审批、审计和系统配置接口。",
        "title": "CertFlow API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/settings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "获取系统基础配置",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.pkiSettings"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Settings"
                ],
                "summary": "保存系统基础配置",
                "parameters": [
                    {
                        "description": "系统基础配置",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/router.pkiSettings"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.pkiSettings"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.errorDocResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "router.errorDocResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "router.pkiSettings": {
            "type": "object",
            "properties": {
                "appName": {
                    "type": "string"
                },
                "appSubtitle": {
                    "type": "string"
                },
                "autoRenew": {
                    "type": "boolean"
                },
                "crlEnabled": {
                    "type": "boolean"
                },
                "crlIntervalHours": {
                    "type": "integer"
                },
                "crlUrl": {
                    "type": "string"
                },
                "expiryNotificationCron": {
                    "type": "string"
                },
                "expiryNotificationDays": {
                    "type": "integer"
                },
                "iconData": {
                    "type": "string"
                },
                "loginName": {
                    "type": "string"
                },
                "ocspEnabled": {
                    "type": "boolean"
                },
                "ocspUrl": {
                    "type": "string"
                },
                "passwordResetRateLimitMinutes": {
                    "type": "integer"
                },
                "passwordResetSendCooldownMinutes": {
                    "type": "number"
                },
                "renewDays": {
                    "type": "integer"
                },
                "resetCaptchaTtlMinutes": {
                    "type": "integer"
                },
                "resetCodeTtlMinutes": {
                    "type": "integer"
                },
                "siteName": {
                    "type": "string"
                },
                "wecomStateTtlMinutes": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}